Prevent conversion of numeric literals to Number type
The problem solved here can be demonstrated by evaluating
`9999999999999999.0 - 9999999999999998.0`. [0]  The correct answer is 1;
SelectCalc previously gave 2 despite using decimal math for operators.
This was caused by expr-eval.js applying the built-in parseFloat() and
parseInt() functions to numeric literals (increasing the left-hand side
by 1 in this example) and is "fixed" by temporarily monkey patching
these functions while the expression is parsed.  This must be done in
the background script so as to not interfere with Web content.

[0]: <https://geocar.sdf1.org/numbers.html>

           I TRIED TO PARSE A DECIMAL WITH JAVASCRIPT

                                ,--.
                               /-__ \
                              |\\\   |
                              |\\\\/|D
                              = \  //
                          __=--`-\_\--,__
                         /#######\   \###`\
                        /         \   \|  |
                       /   /|      \   \  |
                      /   / |       \     |
                     /   /  |/u/surkh\    /

               I NOW HAVE A  DECIMALPARSERFACTORY